Feature/actions (#17)
* DB Message Actions part

* DB Message Actions part

* Message Actions part 2,
Repository, mappers, services added

* Renaming Reactions to Actions

* Styling, missing docs

* Restored reaction code

* ReactionViewModel extracted from MessageViewModel

* New Reactions UI

* Message item onClick removed,
Theme Defaults extracted,

* Theme Defaults fix

* Theme Defaults fix

* MessageAction ID aligned with iOS

* Dao add / update methods merged to insertOrUpdate

* Error Handler added

* Review changes

* DBMessageAction with UserId

* Reaction Count higher than 99 will show 99+

* Merge with develop

* New mapper with ID added,
Theme fixed,
Align fetch methods,

* New mapper with ID added,
Theme fixed,
Align fetch methods,
Code cleanup

* Merge with develop
