+++ chat-components/framework/src/main/java/com/pubnub/framework/mapper/Mapper.kt
@@ -6,3 +6,8 @@ package com.pubnub.framework.mapper
 interface Mapper<in Input, out Output> {
     fun map(input: Input): Output
 }
+
+interface MapperWithId<in Input, out Output> {
+    fun map(id: String, input: Input): Output
+}
+
