+++ chat-components/src/main/java/com/pubnub/components/chat/network/mapper/NetworkMessageActionHistoryMapper.kt
@@ -3,17 +3,15 @@ package com.pubnub.components.chat.network.mapper
 import com.pubnub.api.models.consumer.history.PNFetchMessageItem
 import com.pubnub.components.data.message.action.DBMessageAction
 import com.pubnub.framework.data.ChannelId
-import com.pubnub.framework.mapper.Mapper
+import com.pubnub.framework.mapper.MapperWithId
 
-class NetworkMessageActionHistoryMapper : Mapper<PNFetchMessageItem, Array<DBMessageAction>> {
-    lateinit var channel: ChannelId
-
-    override fun map(input: PNFetchMessageItem): Array<DBMessageAction> {
+class NetworkMessageActionHistoryMapper : MapperWithId<PNFetchMessageItem, Array<DBMessageAction>> {
+    override fun map(id: ChannelId, input: PNFetchMessageItem): Array<DBMessageAction> {
         return input.actions?.flatMap { (type, list) ->
             list.flatMap { (action, users) ->
                 users.map { user ->
                     DBMessageAction(
-                        channel = channel,
+                        channel = id,
                         user = user.uuid,
                         messageTimestamp = input.timetoken,
                         published = user.actionTimetoken.toLong(),
