+++ chat-components/src/main/java/com/pubnub/components/chat/network/mapper/NetworkMessageHistoryMapper.kt
@@ -9,13 +9,12 @@ import com.pubnub.api.models.consumer.history.PNFetchMessageItem
 import com.pubnub.components.chat.network.data.NetworkMessage
 import com.pubnub.components.data.message.DBMessage
 import com.pubnub.framework.data.ChannelId
-import com.pubnub.framework.mapper.Mapper
+import com.pubnub.framework.mapper.MapperWithId
 import com.pubnub.framework.util.toJson
 import java.lang.reflect.Type
 
 class NetworkMessageHistoryMapper(private val mapper: MapperManager) :
-    Mapper<PNFetchMessageItem, DBMessage> {
-    lateinit var channel: ChannelId
+    MapperWithId<PNFetchMessageItem, DBMessage> {
 
     private val gson = GsonBuilder()
         .registerTypeAdapter(NetworkMessage.Attachment::class.java, attachmentDeserializer())
@@ -23,7 +22,7 @@ class NetworkMessageHistoryMapper(private val mapper: MapperManager) :
         .create()
 
     @Suppress("UNCHECKED_CAST")
-    override fun map(input: PNFetchMessageItem): DBMessage {
+    override fun map(id: ChannelId, input: PNFetchMessageItem): DBMessage {
         val message: NetworkMessage =
             gson.fromJson(input.message.toJson(mapper), NetworkMessage::class.java)
         return DBMessage(
@@ -33,7 +32,7 @@ class NetworkMessageHistoryMapper(private val mapper: MapperManager) :
             attachment = message.attachment.toDb(),
             custom = message.custom as Map<String, Any>?,
             publisher = input.uuid!!,
-            channel = channel,
+            channel = id,
             timetoken = input.timetoken,
             isSent = true,
             exception = null,
