+++ chat-components/src/main/java/com/pubnub/components/chat/network/paging/MessageRemoteMediator.kt
@@ -122,7 +122,7 @@ class MessageRemoteMediator constructor(
 
     private suspend fun loadNewMessages(channelId: ChannelId, start: Long?, end: Long?) {
         // Get the history
-        service.pullHistory(
+        service.fetchAll(
             id = channelId,
             start = start,
             end = end,
