+++ chat-components/ui/src/main/java/com/pubnub/components/chat/ui/component/common/TextTheme.kt
@@ -29,7 +29,6 @@ class TextTheme(
     overflow: TextOverflow,
     softWrap: Boolean,
     maxLines: Int,
-    // TODO: 5/25/21 remove text style!
     style: TextStyle,
 ) {
     var modifier by mutableStateOf(modifier, structuralEqualityPolicy())
